refactor: use same type in components
